# api/openapi.yaml
openapi: 3.1.0
info:
  title: ILTER Gateway API
  description: Secure, cost-optimizing LLM Gateway with built-in PII masking, loop detection, and load balancing.
  version: 1.0.0
servers:
  - url: http://127.0.0.1:9191
    description: Dashboard Server
paths:
...
  /authorize:
    get:
      tags: [OAuth 2.0]
      servers:
        - url: http://127.0.0.1:8181
      summary: OAuth Authorization Page (redirect)
      description: |
        OAuth authorization endpoint. Validates PKCE parameters then redirects the
        browser to the dashboard consent page. This is step 4 of the MCP HTTP auth
        flow — the MCP client opens this URL in the user's browser.
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
          description: Must be "code"
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: Client identifier
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
          description: Redirect URI (must be a loopback address, localhost, or vscode://)
          example: http://127.0.0.1:34567
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
          description: PKCE code challenge method (only S256 supported)
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
          description: PKCE code challenge (SHA-256 hash of code verifier, base64url-encoded)
        - name: state
          in: query
          required: false
          schema:
            type: string
          description: Opaque state value for CSRF protection (recommended)
      responses:
        '302':
          description: Redirect to the dashboard consent page
          headers:
            Location:
              schema:
                type: string
              description: URL of the dashboard /authorize page with authorization params
        '400':
          $ref: '#/components/responses/OAuthError'
        '405':
          $ref: '#/components/responses/OAuthError'

    post:
      tags: [OAuth 2.0]
      servers:
        - url: http://127.0.0.1:8181
      summary: OAuth Authorization Consent
      description: |
        Called by the dashboard UI after user consent. Issues an opaque
        authorization code bound to an API key. Returns a redirect_uri so
        the dashboard JS can redirect the browser to the MCP client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - request_id
                - action
              properties:
                request_id:
                  type: string
                  description: Pending authorization request ID (from the GET redirect)
                action:
                  type: string
                  enum: [create_new, use_existing]
                  description: Whether to create a new API key or use an existing one
                api_key:
                  type: string
                  description: Existing API key (required when action=use_existing)
      responses:
        '200':
          description: Authorization code issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  redirect_uri:
                    type: string
                    description: Full redirect URI including code and state query params
                  code:
                    type: string
                    description: Opaque single-use authorization code
                  state:
                    type: string
                    description: State value echoed back for CSRF verification
        '400':
          $ref: '#/components/responses/OAuthError'
        '405':
          $ref: '#/components/responses/OAuthError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  responses:
...
    OAuthError:
      description: OAuth error response (RFC 6749 Section 5.2)
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                description: OAuth error code
                example: invalid_request
              error_description:
                type: string
                description: Human-readable error description
                example: client_id is required